make register controller, fix cloudinary, user model

+++ src/utils/cloudinary.js
@@ -1,25 +1,34 @@
 import { v2 as cloudinary } from 'cloudinary';
 import fs from "fs"
-
-// Configuration
-cloudinary.config({
-  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
-  api_key: process.env.CLOUDINARY_API_KEY,
-  api_secret: process.env.CLOUDINARY_API_SECRET
-});
+import path from 'path';
 
 const uploadOnCloudinary = async (localFilePath) => {
   try {
     if (!localFilePath) return null
+    // Configuration
+    cloudinary.config({
+      cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+      api_key: process.env.CLOUDINARY_API_KEY,
+      api_secret: process.env.CLOUDINARY_API_SECRET
+    });
+    console.log("Cloudinary config:", {
+      cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+      api_key: process.env.CLOUDINARY_API_KEY,
+    })
     // upload the file on cloudinary
-    const response = await cloudinary.uploader.upload(localFilePath, {
+    const absolutePath = path.resolve(localFilePath)
+    const response = await cloudinary.uploader.upload(absolutePath, {
       resource_type: "auto"
     })
     // file has been uploaded successfully
+    fs.unlinkSync(absolutePath)
     console.log("file is uploaded on cloudinary", response.url);
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
+    if (fs.existsSync(localFilePath)) {
+      console.log("Cloudinary upload error:", error) 
+      fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
+    }
     return null;
   }
 }
